+++ hutool-extra/src/test/java/cn/hutool/extra/pinyin/PinyinUtilTest.java
@@ -1,6 +1,5 @@
 package cn.hutool.extra.pinyin;
 
-import cn.hutool.core.util.ArrayUtil;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -8,19 +7,19 @@ public class PinyinUtilTest {
 
 	@Test
 	public void getPinyinTest(){
-		final String pinyin = PinyinUtil.getPinyin("你好", false);
+		final String pinyin = PinyinUtil.getPinyin("你好", " ");
 		Assert.assertEquals("ni hao", pinyin);
 	}
 
 	@Test
 	public void getPinyinUpperCaseTest(){
-		final String pinyin = PinyinUtil.getPinyin("你好怡", true);
-		Assert.assertEquals("NI HAO YI", pinyin);
+		final String pinyin = PinyinUtil.getPinyin("你好怡", " ");
+		Assert.assertEquals("ni hao yi", pinyin);
 	}
 
 	@Test
 	public void getFirstLetterTest(){
-		final char[] result = PinyinUtil.getFirstLetter("H是第一个", false);
-		Assert.assertEquals("h, s, d, y, g", ArrayUtil.join(result, ", "));
+		final String result = PinyinUtil.getFirstLetter("H是第一个", ", ");
+		Assert.assertEquals("h, s, d, y, g", result);
 	}
 }