+++ hutool-json/src/main/java/cn/hutool/json/JSONUtil.java
@@ -335,6 +335,7 @@ public final class JSONUtil {
 	 * 转为JSON字符串，并写出到write
 	 *
 	 * @param json JSON
+	 * @param writer Writer
 	 * @since 5.3.3
 	 */
 	public static void toJsonStr(JSON json, Writer writer) {